MOVE: QR 생성 컴포넌트 util로 이동
